+++ libs/feature/lesson/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./lib/lesson-layout";
 export * from "./lib/use-lesson-context";
 export * from "./lib/license-view-modal";
-export * from "./lib/lesson-utils";
+export * from "./lib/lesson-data-access";
